FeedRankApi/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
ptvsd==4.3.2 # Required for debugging.
djangorestframework
markdown
django-filter
tensorflow
keras
pygeohash
pandas
matplotlib
sklearn
pillow

